fix : controller에서 많은 역할을 하는것을 수정
- controller에서 레이싱카에 대한 일을 하는 거 같아 이것을 racingcar class하도록 수정
- var을 public하게 쓰는것보다 private으로 쓰는게 좋음
