+++ src/main/kotlin/racingcar/model/RacingCar.kt
@@ -1,3 +1,22 @@
 package racingcar.model
 
-data class RacingCar(val racingCarName: String, var distance: Int)
+import camp.nextstep.edu.missionutils.Randoms
+
+data class RacingCar(val racingCarName: String, private var distance: Int) {
+    fun getDistance(): Int {
+        return distance
+    }
+
+    fun addRacingCarDistance() {
+        val randomNum = Randoms.pickNumberInRange(RANDOM_START_NUM, RANDOM_FINISH_NUM)
+        if (randomNum >= MIN_DISTANCE_NUM) {
+            distance += randomNum
+        }
+    }
+
+    companion object {
+        private const val RANDOM_START_NUM = 0
+        private const val RANDOM_FINISH_NUM = 9
+        const val MIN_DISTANCE_NUM = 4
+    }
+}
