+++ src/main/kotlin/racingcar/view/InputView.kt
@@ -30,9 +30,9 @@ class InputView {
 
     companion object {
         private const val NAME_DELIMITERS = ","
-        const val INIT_DISTANCE_NUM = 0
         private const val INPUT_CAR_NAME_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)"
         private const val INPUT_GAME_COUNT_MESSAGE = "시도할 횟수는 몇 회인가요?"
         const val MAX_NAME_LENGTH_NUM = 5
+        const val INIT_DISTANCE_NUM = 0
     }
 }
